requirements.txt
matplotlib
tensorflow
keras
Pillow
opencv-python
imutils
scikit-learn